Fix separable gaussian

+++ src/graphics_opengl.cpp
@@ -543,7 +543,7 @@ int Graphics::LoadTexture(int width, int height, int components, int format, voi
 
 	if (anisotropic)
 	{
-		glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_MAX_ANISOTROPY_EXT, anisotropic);
+		glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_MAX_ANISOTROPY_EXT, (float)anisotropic);
 	}
 
 	glTexImage2D(GL_TEXTURE_2D, 0, components, width, height, 0, format, GL_UNSIGNED_BYTE, bytes);
